% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonn-package.R
\docType{package}
\name{bonn-package}
\alias{bonn-package}
\title{Data retrieval client for the INKAR Database}
\description{
Data retrieval client for the INKAR Database
}
\author{
Moritz Marbach \email{moritz.marbach@tamu.edu}
}
