% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geographies.R
\name{get_geographies}
\alias{get_geographies}
\title{Get Available Geographies}
\usage{
get_geographies(geography = NULL)
}
\arguments{
\item{geography}{geography as string or \code{NULL}}
}
\value{
a \code{data.frame} either with the columns \code{Kurzname} (geography name), \code{ID} (geography identifier) 
and \code{NGeb} (number of units) or with the columns \code{Schlüssel} (unit identifier) and \code{Name} (unit name).
}
\description{
\code{get_geographies} either retrieves a data frame with all available geographies or all existing units for a given geography
}
\examples{

 # Available geographies  
 get_geographies()

	# All districts 
	get_geographies("KRS")



}
